+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/services/ConsumerService.java
@@ -23,6 +23,8 @@ import java.util.function.Predicate;
 
 import lombok.RequiredArgsConstructor;
 
+import org.springframework.security.access.prepost.PostAuthorize;
+import org.springframework.security.access.prepost.PostFilter;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
 
@@ -73,10 +75,16 @@ public class ConsumerService {
     return Optional.ofNullable(consumer);
   }
 
+  @PostAuthorize("returnObject.isEmpty() ? true: hasPermission(returnObject, 'READ')")
+  public Optional<Consumer> get(ConsumerKey key) {
+    return read(key);
+  }
+
   public Optional<Consumer> read(ConsumerKey key) {
     return consumerRepository.findById(key);
   }
 
+  @PostFilter("hasPermission(filterObject, 'READ')")
   public List<Consumer> findAll(Predicate<Consumer> filter) {
     return consumerRepository.findAll().stream().filter(filter).collect(toList());
   }