+++ graphql/api/src/main/java/com/expediagroup/streamplatform/streamregistry/graphql/query/impl/InfrastructureQueryImpl.java
@@ -36,7 +36,7 @@ public class InfrastructureQueryImpl implements InfrastructureQuery {
 
   @Override
   public Optional<Infrastructure> byKey(InfrastructureKeyInput key) {
-    return infrastructureService.read(key.asInfrastructureKey());
+    return infrastructureService.get(key.asInfrastructureKey());
   }
 
   @Override