+++ graphql/api/src/main/java/com/expediagroup/streamplatform/streamregistry/graphql/query/impl/ZoneQueryImpl.java
@@ -36,7 +36,7 @@ public class ZoneQueryImpl implements ZoneQuery {
 
   @Override
   public Optional<Zone> byKey(ZoneKeyInput key) {
-    return zoneService.read(key.asZoneKey());
+    return zoneService.get(key.asZoneKey());
   }
 
   @Override