+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## UNRELEASED
+### Added
+- Authorisation for read, findAll, and find operations for Consumer, ConsumerBinding, 
+  Domain, Infrastructure, Producer, ProducerBinding, Schema, Stream, StreamBinding, Zone
+  
 ### Changed
 - Refactored upsert calls in mutations and removed upsert methods from services
-- Refactored update_status calls 
+- Refactored update_status calls
 
 ## [0.14.5] - 2020-10-19
 ### Added